Finished

+++ ft_printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: dacaball <dacaball@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/02 13:23:16 by dacaball          #+#    #+#             */
-/*   Updated: 2023/10/06 11:33:35 by dacaball         ###   ########.fr       */
+/*   Updated: 2023/10/09 11:08:17 by dacaball         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,10 +20,10 @@ int		ft_printf(char const *str, ...);
 
 void	ft_printstr(char *str, int *i);
 void	ft_printchar_fd(char chr, int *i);
-void	ft_printhexlow(size_t decimal, int *i);
-void	ft_printhexup(size_t decimal, int *i);
+void	ft_printhexlow(long long unsigned decimal, int *i);
+void	ft_printhexup(long long unsigned decimal, int *i);
 void	ft_printptr(size_t	ptr, int *i);
-void	ft_printdec(long int decimal, int *i);
+void	ft_printdec(long long int decimal, int *i);
 void	ft_printunsigned(unsigned int decimal, int *i);
 
 #endif
