+++ ft_printf/ft_printhexlow.c
@@ -6,13 +6,13 @@
 /*   By: dacaball <dacaball@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/04 22:30:07 by dacaball          #+#    #+#             */
-/*   Updated: 2023/10/05 11:40:41 by dacaball         ###   ########.fr       */
+/*   Updated: 2023/10/09 11:08:54 by dacaball         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-void	ft_printhexlow(size_t decimal, int *i)
+void	ft_printhexlow(long long unsigned decimal, int *i)
 {
 	char const	base_hex[] = "0123456789abcdef";
