+++ ft_printf/ft_printhexup.c
@@ -6,13 +6,13 @@
 /*   By: dacaball <dacaball@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/04 13:31:14 by dacaball          #+#    #+#             */
-/*   Updated: 2023/10/05 11:41:14 by dacaball         ###   ########.fr       */
+/*   Updated: 2023/10/09 11:07:54 by dacaball         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-void	ft_printhexup(size_t decimal, int *i)
+void	ft_printhexup(long long unsigned decimal, int *i)
 {
 	char const	base_hex[] = "0123456789ABCDEF";
